tests/wpt/tests/css/css-box/margin-trim/computed-margin-values/flexbox-column-multi-line-inline-end.html

<!DOCTYPE html>
<html>
<head>
<link rel="author" href="mailto:sammy.gill@apple.com">
<link rel="help" href="https://drafts.csswg.org/css-box-4/#margin-trim-flex">
<meta name="assert" content="trimmed inline-end margins should be reflected in computed style">
<style>
flexbox {
    display: flex;
    flex-direction: column;
    flex-wrap: wrap;
    width: min-content;
    height: 100px;
    margin-trim: inline-end;
}
item {
    display: block;
    background-color: green;
    width: 50px;
    height: 50px;
    margin-inline-end: 10px;
}
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
</head>
<body onload="checkLayout('flexbox > item')">
<div id="target">
<flexbox>
    <item data-offset-x="8" data-expected-margin-right="10"></item>
    <item data-offset-x="8" data-expected-margin-right="10"></item>
    <item data-offset-x="68" data-expected-margin-right="0"></item>
    <item data-offset-x="68" data-expected-margin-right="0"></item>
</flexbox>
</div>
</body>
</html>
