third_party/blink/web_tests/external/wpt/css/css-box/margin-trim/computed-margin-values/flexbox-column-multi-line-inline-end-expected.txt

This is a testharness.js-based test.
[FAIL] flexbox > item 3
  assert_equals: \n<item data-offset-x="68" data-expected-margin-right="0"></item>\nmargin-right expected "0" but got "10"
[FAIL] flexbox > item 4
  assert_equals: \n<item data-offset-x="68" data-expected-margin-right="0"></item>\nmargin-right expected "0" but got "10"
Harness: the test ran to completion.
